# api_gateway/docs/swagger.yaml
basePath: /
definitions:
  dto.NewOrderDetailingPackageItems:
    properties:
      detailing_package_id:
        type: integer
      qty:
        type: integer
    required:
    - detailing_package_id
    type: object
  dto.NewOrderRequest:
    properties:
      address:
        type: string
      detailing_package_items:
        items:
          $ref: '#/definitions/dto.NewOrderDetailingPackageItems'
        type: array
      wash_package_items:
        items:
          $ref: '#/definitions/dto.NewOrderWashPackageItems'
        type: array
    required:
    - address
    - detailing_package_items
    - wash_package_items
    type: object
  dto.NewOrderWashPackageItems:
    properties:
      qty:
        type: integer
      wash_package_id:
        type: integer
    required:
    - wash_package_id
    type: object
...
  dto.SwaggerResponseOrder:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.Order'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  dto.SwaggerResponseUpdateOrderStatus:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.Order'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
  dto.SwaggerResponseWasherGetAllOrders:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.Orders'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
  dto.SwaggerResponseWasherGetCurrentOrder:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/orderpb.Order'
        x-order: "1"
      message:
        type: string
        x-order: "0"
    type: object
...
  orderpb.Address:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  orderpb.DetailingPackage:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      qty:
        type: integer
      subtotal:
        type: number
    type: object
  orderpb.Order:
    properties:
      address:
        $ref: '#/definitions/orderpb.Address'
      created_at:
        type: string
      object_id:
        type: string
      order_detail:
        $ref: '#/definitions/orderpb.OrderDetail'
      payment:
        $ref: '#/definitions/orderpb.Payment'
      status:
        type: string
      user:
        $ref: '#/definitions/orderpb.User'
      washer:
        $ref: '#/definitions/orderpb.Washer'
    type: object
  orderpb.OrderDetail:
    properties:
      app_fee:
        type: number
      detailing_packages:
        items:
          $ref: '#/definitions/orderpb.DetailingPackage'
        type: array
      total_price:
        type: number
      wash_packages:
        items:
          $ref: '#/definitions/orderpb.WashPackage'
        type: array
    type: object
  orderpb.Orders:
    properties:
      orders:
        items:
          $ref: '#/definitions/orderpb.Order'
        type: array
    type: object
  orderpb.Payment:
    properties:
      invoice_id:
        type: string
      invoice_url:
        type: string
      method:
        type: string
      status:
        type: string
      total:
        type: number
    type: object
  orderpb.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  orderpb.WashPackage:
    properties:
      category:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      qty:
        type: integer
      subtotal:
        type: number
    type: object
  orderpb.Washer:
    properties:
      id:
        type: integer
      name:
        type: string
      status:
        type: string
    type: object
  utils.ErrResponse:
    properties:
      detail:
        type: string
        x-order: "2"
      status:
        type: integer
        x-order: "0"
      type:
        type: string
        x-order: "1"
    type: object
...
host: 3ee9-103-18-34-211.ngrok-free.app
info:
  contact:
    email: muhlisiqbalutomoo@gmail.com
    name: The developer
  description: Online motorcycle washing service built with microservices that integrates
    user, washer, and admin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Fox Wash
  version: 0.0.1
paths:
...
  /users/orders:
    post:
      consumes:
      - application/json
      description: Create a new order for the logged in user. You will need an 'Authorization'
        cookie attached with this request.
      parameters:
      - description: Order details
        in: body
        name: orderRequest
        required: true
        schema:
          $ref: '#/definitions/dto.NewOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SwaggerResponseOrder'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Create new order
      tags:
      - customer
...
  /washers/orders:
    get:
      description: Retrieve all orders related to the logged in washer. You will need
        an 'Authorization' cookie attached with this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseWasherGetAllOrders'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get all washer's orders
      tags:
      - washer
  /washers/orders/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve an order by it's ID. You will need an 'Authorization'
        cookie attached with this request.
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseOrder'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get order by ID
      tags:
      - washer
  /washers/orders/ongoing:
    get:
      description: Retrieve ongoing order related to the logged in washer. You will
        need an 'Authorization' cookie attached with this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseWasherGetCurrentOrder'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Get current washer's order
      tags:
      - washer
  /washers/orders/status/{id}:
    put:
      consumes:
      - application/json
      description: Updates ongoing order status related to the logged in washer. You
        will need an 'Authorization' cookie attached with this request.
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SwaggerResponseUpdateOrderStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrResponse'
      summary: Update washer's order
      tags:
      - washer
...
swagger: "2.0"